move routes outside of spotify
Created class for routes specifically for the data they serve
